requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
boto3==1.34.157
botocore==1.34.157
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
infobip-api-python-client==4.0.0
jmespath==1.0.1
kombu==5.3.7
multidict==6.0.5
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
